Refactor new landing page into components and other cleanup.
