+++ modules/odr_frontend/src/lib/landing_page/CommunityReviewTab.svelte
@@ -1,0 +1,8 @@
+<!--
+  SPDX-License-Identifier: Apache-2.0
+-->
+
+<!-- Community Review content will go here -->
+<div class="text-center p-8">
+  Community Review content
+</div>
