+++ modules/odr_frontend/src/lib/landing_page/MyContributionsTab.svelte
@@ -1,0 +1,8 @@
+<!--
+  SPDX-License-Identifier: Apache-2.0
+-->
+
+<!-- My Contributions content will go here -->
+<div class="text-center p-8">
+  My Contributions content
+</div>
